requirements.txt
asttokens==2.0.5
backcall==0.2.0
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fonttools==4.34.4
future==0.18.2
imutils==0.5.4
ipython==8.4.0
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
stack-data==0.3.0
tqdm==4.64.0
traitlets==5.3.0
wcwidth==0.2.5
